requirements.txt
openai>=1.0
chromadb>=0.4
pandas>=2.0
python-dotenv>=1.0
pytest>=7.0
